requirements.txt
python-deriv-api
pandas
numpy
ta
streamlit
TA-Lib
scikit-learn
tensorflow